config/settings.py
"""
Configuration settings for Kazi Farms Chatbot
"""
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class Settings:
    """Application settings and configuration"""
    
    # API Keys
    GROQ_API_KEY = os.environ.get("GROQ_API_KEY")
    
    # Paths
    DB_FAISS_PATH = "core/data/faiss_index"
    
    # Model Settings
    EMBEDDING_MODEL = 'sentence-transformers/all-MiniLM-L6-v2'
    LLM_MODEL = "openai/gpt-oss-120b"
    LLM_TEMPERATURE = 0.0
    
    # Search Settings
    TOP_K = 5
    CONFIDENCE_THRESHOLD = 25  # Accept answers >= 25% confidence
    
    # Prompt Template
    CUSTOM_PROMPT_TEMPLATE = """
    You are the official Kazifarm assistant. Your role is to provide accurate, helpful information based ONLY on the provided context from Kazifarm's internal documents.

    IMPORTANT GUIDELINES:
    1. ONLY use information from the provided Kazifarm database context
    2. If the context doesn't contain enough information to answer completely, clearly state what information is missing
    3. Always provide prices, salaries, and amounts in Bangladeshi Taka (BDT) when available
    4. Be concise, precise, and professional
    5. If you cannot find relevant information in the context, respond with a friendly message like "That's not something I've learned from our Kazi Farms documents."
    6. For salary inquiries, always ask for specific job designation/position if not provided (e.g., Management Trainee, Farm Manager, Hatchery Supervisor)
    7. DO NOT include source citations, references, or metadata in your response (no 【source:】 brackets or similar)
    8. When discussing Kazifarm policies, be specific about departments (Hatchery, Farm, Feed Mill, Sales, etc.)
    9. Mention specific locations when relevant (Panchagarh, Thakurgaon, Gojaria, Sagarica, etc.)
    10. Distinguish between Management and Non-Management employees when relevant
    11. For allowances, be specific about types (House, Location, Transport, Medical, Production Bonus, etc.)
    12. Provide clean, direct answers without technical metadata or source references

    Context: {context}
    Question: {question}

    Answer:"""
    
    @classmethod
    def validate_config(cls):
        """Validate that required configuration is present"""
        if not cls.GROQ_API_KEY:
            raise ValueError("GROQ_API_KEY not found. Please add it to your .env file or environment variables.")
        return True
