requirements.txt
langchain
langchain-core
langchain-huggingface
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
streamlit
